refactor signify.ts
