+++ src/keri/app/exchanging.ts
@@ -1,19 +1,101 @@
-import {b, Dict, Ident, Ilks, Serials, versify} from "../core/core";
-import {Serder} from "../core/serder";
-import {nowUTC} from "../core/utils";
-import {Pather} from "../core/pather";
-import {Counter, CtrDex} from "../core/counter";
-import {Saider} from "../core/saider";
+import { SignifyClient } from "./clienting"
+import { b, d, Dict, Ident, Ilks, Serials, versify } from "../core/core";
+import { Serder } from "../core/serder";
+import { nowUTC } from "../core/utils";
+import { Pather } from "../core/pather";
+import { Counter, CtrDex } from "../core/counter";
+import { Saider } from "../core/saider";
+
+/**
+ * Exchanges
+ */
+export class Exchanges {
+    client: SignifyClient
+
+    /**
+     * Exchanges
+     * @param {SignifyClient} client
+     */
+    constructor(client: SignifyClient) {
+        this.client = client
+    }
+
+    /**
+     * Create exn message
+     * @async
+     * @returns {Promise<any>} A promise to the list of replay messages
+     * @param sender
+     * @param route
+     * @param payload
+     * @param embeds
+     */
+    createExchangeMessage(sender: Dict<any>, route: string, payload: Dict<any>, embeds: Dict<any>): [Serder, string[], string] {
+        let keeper = this.client.manager!.get(sender)
+        let [exn, end] = exchange(route,
+            payload,
+            sender["prefix"], undefined, undefined, undefined, undefined, embeds)
+
+        let sigs = keeper.sign(b(exn.raw))
+        return [exn, sigs, d(end)]
+    }
+
+    /**
+     * Send exn messages to list of recipients
+     * @async
+     * @returns {Promise<any>} A promise to the list of replay messages
+     * @param name
+     * @param topic
+     * @param sender
+     * @param route
+     * @param payload
+     * @param embeds
+     * @param recipients
+     */
+    async send(name: string, topic: string, sender: Dict<any>, route: string, payload: Dict<any>, embeds: Dict<any>,
+        recipients: string[]): Promise<any> {
+
+        let [exn, sigs, atc] = this.createExchangeMessage(sender, route, payload, embeds)
+        return await this.sendFromEvents(name, topic, exn, sigs, atc, recipients)
+
+    }
+
+    /**
+     * Send exn messaget to list of recipients
+     * @async
+     * @returns {Promise<any>} A promise to the list of replay messages
+     * @param name
+     * @param topic
+     * @param exn
+     * @param sigs
+     * @param atc
+     * @param recipients
+     */
+    async sendFromEvents(name: string, topic: string, exn: Serder, sigs: string[], atc: string, recipients: string[]): Promise<any> {
+
+        let path = `/identifiers/${name}/exchanges`
+        let method = 'POST'
+        let data: any = {
+            tpc: topic,
+            exn: exn.ked,
+            sigs: sigs,
+            atc: atc,
+            rec: recipients
+        }
+
+        let res = await this.client.fetch(path, method, data)
+        return await res.json()
+    }
+}
 
 
 export function exchange(route: string,
-                         payload: Dict<any>,
-                         sender: string,
-                         recipient?: string,
-                         date?: string,
-                         dig?: string,
-                         modifiers?: Dict<any>,
-                         embeds?: Dict<any>): [Serder, Uint8Array] {
+    payload: Dict<any>,
+    sender: string,
+    recipient?: string,
+    date?: string,
+    dig?: string,
+    modifiers?: Dict<any>,
+    embeds?: Dict<any>): [Serder, Uint8Array] {
 
 
     const vs = versify(Ident.KERI, undefined, Serials.JSON, 0)
